Vector2ScrubOverlay component
